Completed exercise 7, holy grail website
Website with header, footer, sidebar, then conten on the left.
Used flexbox to format the whole website. Trickiest part was the
content on the right side of the page. Since it is 6 boxes with gaps
in between them, I had to think for a little bit.
So I created a flex container that had the sidebar and content.
This allowed for auto-margin on the bottom with the footer.
To format the content, I created a flex container with only content.
Ie the stuff right of the sidebar, above the footer below the header.
Then we needed flexwrap for the cards to wrap and take the whole space.
viola.